# docs/swagger/swagger.yaml
basePath: /api
definitions:
  models.AgeGuess:
    properties:
      age:
        type: integer
    type: object
  models.AuthenticationResponse:
    properties:
      token:
        type: string
      userId:
        type: string
    type: object
  models.Credentials:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.GuessResponse:
    properties:
      correct:
        type: boolean
      message:
        type: string
    type: object
  models.Image:
    properties:
      age:
        type: integer
      createdAt:
        type: string
      id:
        type: string
      url:
        type: string
      userId:
        type: string
    type: object
  models.ImagesResponse:
    properties:
      images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
    type: object
  models.PersonalDataResponse:
    properties:
      age:
        type: integer
      email:
        type: string
      gender:
        type: string
      image:
        type: string
      lastName:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
  models.User:
    properties:
      age:
        type: integer
      email:
        type: string
      gender:
        type: string
      id:
        type: string
      image:
        type: string
      lastName:
        type: string
      name:
        type: string
      password:
        type: string
      ratedImages:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
host: localhost:5000
info:
  contact:
    email: luispalominot@hotmail.com
    name: Luis Palomino
  description: This is the API documentation for GuessIt
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GuessIt API
  version: "0.1"
paths:
  /Image/:
    get:
      description: Get all images
      operationId: get-images-endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ImagesResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Retrieve all images. If user is logged in, his/her images are not returned.
  /Image/{id}/Rate:
    post:
      consumes:
      - application/json
      operationId: rate-image-endpoint
      parameters:
      - description: Guess attempt from the user
        in: body
        name: guess
        required: true
        schema:
          $ref: '#/definitions/models.AgeGuess'
          type: object
      - description: ID of the image that needs to be rated
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GuessResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Allows a user to rate an image
  /Image/FromUser:
    get:
      description: Get user images
      operationId: user-images-endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ImagesResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Let a user get its images
  /Image/UploadImage:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image
      operationId: upload-image-endpoint
      parameters:
      - description: Image to be uploaded
        in: formData
        name: image
        required: true
        type: file
      - description: Age that corresponts to the person in the image
        in: formData
        name: age
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Image'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "413":
          description: Request Entity Too Large
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Let a user upload images
  /User/Login:
    post:
      consumes:
      - application/json
      description: login user to system
      operationId: user-login
      parameters:
      - description: User credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.Credentials'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthenticationResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login to system
  /User/Logout:
    post:
      description: Logout user
      operationId: logout-endpoint
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - Bearer: []
      summary: Logout user from system
  /User/PersonalData:
    get:
      description: Retrieve personal data from user
      operationId: personal-data-retrieval
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PersonalDataResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get personal data from a user
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"